Fix FEC build, failing test, and all clippy warnings
- Fix borrow-after-move in FEC stall detection (tcp_transfer.rs) that
  broke compilation under --features fec.
- Replace the never-looping data-listener loop in tcp_server with the
  select! it wrapped (clippy never_loop deny error).
- Update stale calculate_optimal_parallel_streams test to the 64-stream
  auto cap that the implementation, doc comment, and constant all use.
- Make the placeholder timeout-conversion test exercise a real Elapsed.
- Resolve all 42 clippy warnings: clamp patterns, redundant closures,
  io::Error::other, type aliases for complex types, Default impl, etc.

+++ src/integrity.rs
@@ -35,7 +35,7 @@ pub fn hash_file_reader(
     size: u64,
 ) -> Result<[u8; BLAKE3_LEN], TransferError> {
     let mut hasher = blake3::Hasher::new();
-    let mut remaining = size as u64;
+    let mut remaining = size;
     let mut buf = vec![0u8; HASH_CHUNK_SIZE.min(remaining as usize).max(1)];
 
     while remaining > 0 {
